Put category styling in ArticleAlt component
