+++ src/routes/search/+page.svelte
@@ -2,13 +2,9 @@
     import Footer from '$lib/Organism/Footer.svelte';
     import Header from '$lib/Organism/Header.svelte';
     import SearchBar from '$lib/Molecules/SearchBar.svelte';
+    import ArticleAlt from '$lib/Molecules/ArticleAlt.svelte';
 
     export let data;
-
-    const dateFormat = {
-        month: 'short',
-        day: 'numeric',
-    };
 </script>
 
 <Header/>
@@ -23,17 +19,11 @@
     </div>
     {#if data.posts.length > 0}
         <h1>Zoekresultaten voor &quot;{data.searchterm}&quot; :</h1>
-        {#each data.posts as post}
-        <!-- @html means: there is html in this string, render it -->
-            <a href="/{post.slug}">
-                <h3>{@html post.title.rendered}</h3>
-            </a>
-            <p>{@html post.excerpt.rendered}</p>
-            <img src={post.yoast_head_json.og_image[0].url} alt="Artikel afbeelding">
-            <p>{(new Date(post.date)).toLocaleDateString("nl-NL", dateFormat)}</p>
-            <p>{post.yoast_head_json.twitter_misc["Geschatte leestijd"]}</p>
-            <p>{post.yoast_head_json.author}</p>
-        {/each}
+        <div class="posts-container">
+            {#each data.posts as post}
+                <ArticleAlt post={post} />
+            {/each}
+        </div>
     {:else}
         <!-- This will show if no posts are available -->
         <h1>Geen resultaten gevonden voor: &quot;{data.searchterm}&quot;</h1>
@@ -43,14 +33,9 @@
 <Footer/>
 
 <style>
-
     h1 {
         text-align: center;
         margin: 2em;
-    }
-
-    a {
-        color: var(--background-color);
     }
 
     .search-main {
@@ -60,8 +45,12 @@
         margin-right: auto;
     }
 
-    img {
-        max-width: 100%;
-        height: auto;
+    .posts-container {
+        display: flex;
+        flex-direction: column;
+        gap: 1rem;
+        align-items: center;
+        justify-content: center;
+        box-sizing: border-box;
     }
 </style>
